requirements.txt

---

### **requirements.txt**

```plaintext
pandas==2.1.0
requests==2.31.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
psycopg2-binary==2.9.7  # Solo necesario para PostgreSQL
